Update state from db on app reload
