--- src/components/utils/__mocks__/carbonEmission.json
@@ -1,1 +1,0 @@
-{ "carbonFootprint": "16.60" }
